// notusTMDB/src/app/interfaces/videos.interface.ts
// Generated by https://quicktype.io

export interface VideosResponse {
    id:      number;
    results: Videos[];
}

export interface Videos {
    iso_639_1:    string;
    iso_3166_1:   string;
    name:         string;
    key:          string;
    site:         string;
    size:         number;
    type:         string;
    official:     boolean;
    published_at: string;
    id:           string;
}